@@ -22,8 +22,9 @@ using Microsoft.Extensions.DependencyInjection;
 using FluentValidation.AspNetCore;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+//using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using S3.Services.Identity.ExternalEvents;
+using Microsoft.Extensions.Hosting;
 
 namespace S3.Services.Identity
 {
@@ -31,9 +32,9 @@ namespace S3.Services.Identity
     {
         private static readonly string[] Headers = new[] { "X-Operation", "X-Resource", "X-Total-Count" };
         public IConfiguration Configuration { get; }
-        public IContainer Container { get; private set; }
+        //public IContainer Container { get; private set; }
 
-        public Startup(IHostingEnvironment env)
+        public Startup(IWebHostEnvironment env)
         {
             var builder = new ConfigurationBuilder()
                 .SetBasePath(env.ContentRootPath)
@@ -49,8 +50,10 @@ namespace S3.Services.Identity
         //    Configuration = configuration;
         //}
 
-        public IServiceProvider ConfigureServices(IServiceCollection services)
+        public void ConfigureServices(IServiceCollection services)
         {
+            services.AddControllers().AddNewtonsoftJson();
+
             services.AddCustomMvc()
                .AddFluentValidation(fv => fv.RegisterValidatorsFromAssemblyContaining<Startup>()); // Enable fluent validation;
             services.AddSwaggerDocs();
@@ -66,28 +69,47 @@ namespace S3.Services.Identity
                         cors.AllowAnyOrigin()
                             .AllowAnyMethod()
                             .AllowAnyHeader()
-                            .AllowCredentials()
+                            //.AllowCredentials()
                             .WithExposedHeaders(Headers));
             });
            
 
-            var builder = new ContainerBuilder();
+            //var builder = new ContainerBuilder();
+            //builder.RegisterAssemblyTypes(Assembly.GetEntryAssembly())
+            //        .AsImplementedInterfaces();
+            //builder.Populate(services);
+            //builder.AddMongo();
+            //builder.AddMongoRepository<RefreshToken>("RefreshTokens");
+            //builder.AddMongoRepository<User>("Users");
+            //builder.AddRabbitMq();
+            //builder.AddDispatchers();
+            //builder.RegisterType<PasswordHasher<User>>().As<IPasswordHasher<User>>();
+            //Container = builder.Build();
+
+            //return new AutofacServiceProvider(Container);
+        }
+
+        // ConfigureContainer is where you can register things directly
+        // with Autofac. This runs after ConfigureServices so the things
+        // here will override registrations made in ConfigureServices.
+        // Don't build the container; that gets done for you by the factory.
+        public void ConfigureContainer(ContainerBuilder builder)
+        {
+            // Register your own things directly with Autofac, like:
+            //builder.RegisterModule(new MyApplicationModule());
+
             builder.RegisterAssemblyTypes(Assembly.GetEntryAssembly())
-                    .AsImplementedInterfaces();
-            builder.Populate(services);
+                .AsImplementedInterfaces();
             builder.AddMongo();
             builder.AddMongoRepository<RefreshToken>("RefreshTokens");
             builder.AddMongoRepository<User>("Users");
             builder.AddRabbitMq();
             builder.AddDispatchers();
             builder.RegisterType<PasswordHasher<User>>().As<IPasswordHasher<User>>();
-            Container = builder.Build();
-
-            return new AutofacServiceProvider(Container);
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env,
-            IApplicationLifetime applicationLifetime, IConsulClient client,
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env,
+            IHostApplicationLifetime applicationLifetime, IConsulClient client,
             IStartupInitializer startupInitializer)
         {
             if (env.IsDevelopment() || env.EnvironmentName == "local")
@@ -112,7 +134,7 @@ namespace S3.Services.Identity
             applicationLifetime.ApplicationStopped.Register(() =>
             {
                 client.Agent.ServiceDeregister(consulServiceId);
-                Container.Dispose();
+                //Container.Dispose();
             });
 
             startupInitializer.InitializeAsync();
