@@ -1,9 +1,12 @@
-﻿using Microsoft.AspNetCore;
+﻿using Autofac.Extensions.DependencyInjection;
+using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 using S3.Common.Logging;
 using S3.Common.Metrics;
 using S3.Common.Mvc;
 using S3.Common.Vault;
+using System.IO;
 
 namespace S3.Services.Identity
 {
@@ -11,15 +14,28 @@ namespace S3.Services.Identity
     {
         public static void Main(string[] args)
         {
+            // ASP.NET Core 3.0+:
+            // The UseServiceProviderFactory call attaches the
+            // Autofac provider to the generic hosting mechanism.
             CreateWebHostBuilder(args).Build().Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
-                .UseLogging()
-                .UseVault()
-                .UseLockbox()
-                .UseAppMetrics();
+        public static IHostBuilder CreateWebHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .UseServiceProviderFactory(new AutofacServiceProviderFactory())
+                .ConfigureWebHostDefaults(webHostBuilder =>
+                {
+                    webHostBuilder.UseContentRoot(Directory.GetCurrentDirectory())
+                        //.UseIISIntegration()
+                        .UseStartup<Startup>();
+                });
+
+        //public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
+        //    WebHost.CreateDefaultBuilder(args)
+        //        .UseStartup<Startup>()
+        //        .UseLogging()
+        //        .UseVault()
+        //        .UseLockbox()
+        //        .UseAppMetrics();
     }
 }
